登录企业选择、解除关系

+++ cwms-sso/cwms-sso-service/src/main/java/com/lcdt/login/service/AuthTicketService.java
@@ -49,7 +49,7 @@ public class AuthTicketService {
 	}
 
 
-	public boolean generateTicketInResponse(HttpServletRequest request, HttpServletResponse response, Long userId, Integer companyId) {
+	public boolean generateTicketInResponse(HttpServletRequest request, HttpServletResponse response, Long userId, Long companyId) {
 
 		logger.info("generate ticket userId: " + userId + " companyId: " + companyId);
